+++ packages/react/test/react.test.tsx
@@ -1,4 +1,4 @@
-describe('@thefakeorg/react', () => {
+describe('@localfirst/react', () => {
   it('works', () => {
     expect(true).toBe(true);
   });